Data store and querySelector object
A data store and query selector  object created to have a centralised place where all the main data info would be located and executed.

A query Selector object created to have one centralised place where all the the query selector are located. this is to ensure there is no repetition of the query selector calls.